srcs/containers/django/requirements.txt
Django== 4.1 #Django framework
psycopg2== 2.9.9 #PostgreSQL adapter for Python
# psycopg2-binary==2.9.9
gunicorn==21.2.0
django_postgres_extensions== 0.9.3
whitenoise # for serving static files
crispy-bootstrap5 # To integrate Bootstrap 5 with Django forms
requests # making HTTP requests
djangorestframework==3.14.0 # Django REST framework 
djangorestframework-simplejwt # which provides JSON Web Token (JWT) authentication support for Django REST framework.


# djangorestframework-filters
# django-utils-six
# six
# daphne==4.0.0 #This allows you to use WebSockets, long-lived HTTP connections, and other asynchronous protocols in Django.
# channels==4.0.0 #Django Channels allows you to add real-time functionality to your Django applications, such as live notifications, chat applications, and other features that require asynchronous communication.